integrating with google drive/sheets, refactoring code structure
