+++ misc/matching/Pipfile
@@ -12,6 +12,10 @@ pylint = "*"
 click = "*"
 pyyaml = "*"
 psycopg2-binary = "*"
+google-api-python-client = "*"
+google-auth-httplib2 = "*"
+google-auth-oauthlib = "*"
+unidecode = "*"
 
 [requires]
 python_version = "3.7"
